Failed to resolve: com.github.aakumykov:copy_between_streams_with_speed:7.9.3-alpha

+++ app/src/main/java/com/github/aakumykov/sync_dir_to_cloud/domain/entities/SyncTask.kt
@@ -61,7 +61,7 @@ class SyncTask {
     @Deprecated("Не может быть null")
     @ColumnInfo(name = "sync_mode") var syncMode: SyncMode?
 
-    @ColumnInfo(name = "speed", defaultValue = "-1") var speedBytesPerSecond: Long = -1
+    @ColumnInfo(name = "speed", defaultValue = "-1") var speedBytesPerSecond: Int = -1
     @ColumnInfo(name = "parallelism", defaultValue = "1") var parallelism: Int = 1
 
     @ColumnInfo(name = "interval_h") var intervalHours: Int
